Refactorizar capa Business

+++ Vueling.DataAccess.DaoTests/Dao/AlumnoJsonDaoTests.cs
@@ -20,63 +20,49 @@ namespace Vueling.DataAcces.Dao.Tests
     public class AlumnoJsonDaoTests
     {
         private static readonly ILog log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
+        private static readonly string path = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
+        private readonly string fullPath = path + "\\" + "alumnos.json";
         IAlumnoFormatoDao alumnoDao;
         ITypeFactory factory;
 
         [TestInitialize]
         public void testInit()
         {
+            if (File.Exists(fullPath)) File.Delete(fullPath);
+            ConfigUtils.SetValorVarEnvironment("Json");
             factory = new FileFactory();
             alumnoDao = factory.TypeJson();
         }
 
+        [TestCleanup]
+        public void testClean()
+        {
+            File.Delete(fullPath);
+        }
 
-        [DataRow(1, "pere", "zamora", "3333", "01-03-1998")]
+        [DataRow(1, "pere", "zamora", "3333", "01-03-1998","20")]
         [DataTestMethod]
-        public void AddTest(int id, string name, string apellidos, string dni, string fechaNac)
+        public void AddTest(int id, string name, string apellidos, string dni, string fechaNac, string edad)
         {
             log.Debug("Entrar metodo AddTest: ");
-            // Convertir fecha nacimiento en formato DateTime
+
             var lfechaNac = fechaNac.Split('-');
             var FechaNac = new DateTime(Convert.ToInt32(lfechaNac[2]), Convert.ToInt32(lfechaNac[1]), Convert.ToInt32(lfechaNac[0]));
 
-            // Creamos usuario de pruebas (asignamos Guid, fecha creacion y edad calculada)
-            Alumno alumno = new Alumno(id, name, apellidos, dni, FechaNac);
+            Alumno alumno = new Alumno(id, name, apellidos, dni, FechaNac, Convert.ToInt32(edad), DateTime.Now.ToString("yyyyMMddHHmmssffff"));
             alumno.Guid = System.Guid.NewGuid().ToString();
-            alumno.Edad = alumno.CalcularEdat();
-            alumno.FechaCr = alumno.GetTimesTamp(DateTime.Now);
 
-            // Añadimos variable entorno con el formato JSON
-            ConfigUtils.SetValorVarEnvironment("Json");
-
-            // Realizamos la llamada metodo para añadir elemento
             alumnoDao.Add(alumno);
-
-            // Leemos fichero pruebas JSON
             Alumno alumnoTest = LeerAlumnoJson();
 
-            // Comaparamos el alumno insertado con el que hemos recuperado
             Console.WriteLine(alumno);
             Console.WriteLine(alumnoTest);
             Assert.IsTrue(alumno.Equals(alumnoTest));
             log.Debug("Salir metodo AddTest: ");
         }
 
-        [TestCleanup]
-        public void testClean()
-        {
-            String path = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
-            var fullPath = path + "\\" + "alumnos.json";
-
-            File.Delete(fullPath);
-        }
-
         public Alumno LeerAlumnoJson()
         {
-            String path = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
-
-            var fullPath = path + "\\" + "alumnos.json";
-
             // Validar si el insert se ha realizado correctamente
             using (FileStream fs = new FileStream(fullPath, FileMode.Open, FileAccess.Read))
             using (StreamReader sw = new StreamReader(fs))
